chore: Make node sharing lambdas final, try to resolve generic signature of types
With invokedynamic, we are given the generic signature of the
interface method, not the generic signature of the interface.
In general, the generic signature of the interface method cannot
be mapped to the generic signature of the interface class. However,
we can still do it for particular classes which we know how the
method generic signature maps to the class' generic signature.
For instance, a Predicate's type variable(s) are the argument
types in the test method.

Also merged the three input transformer phases into a single
input transformer phase, so if there are other transformers
operating on the ConstraintProvider (for some reason), they
will not encounter the temporarily invalid bytecode generated
by the middle phase.